refactor: Remove modification from access-secret, set annotation on service account
